+++ libs/gmana/lib/either/use_case.dart
@@ -1,0 +1,7 @@
+import 'package:gmana/gmana.dart';
+
+class NoParams {}
+
+abstract interface class UseCase<SuccessType, Params> {
+  FutureEither<SuccessType> call(Params params);
+}