+++ libs/gmana/lib/logger/logger_helper.dart
@@ -1,0 +1,51 @@
+import 'package:flutter/foundation.dart';
+import 'package:logger/logger.dart';
+
+final logger = LoggerHelper();
+
+class LoggerHelper {
+  static final LoggerHelper _instance = LoggerHelper._internal();
+  late Logger _logger;
+
+  factory LoggerHelper() {
+    return _instance;
+  }
+
+  LoggerHelper._internal() {
+    _logger = Logger(
+      printer: PrettyPrinter(
+        methodCount: 0,
+        errorMethodCount: 8,
+        lineLength: 120,
+        colors: true,
+        printEmojis: true,
+        dateTimeFormat: DateTimeFormat.dateAndTime,
+      ),
+      level: kReleaseMode ? Level.off : Level.trace,
+    );
+  }
+
+  void debug(dynamic message, [dynamic error, StackTrace? stackTrace]) {
+    _logger.d(message, error: error, stackTrace: stackTrace);
+  }
+
+  void error(dynamic message, [dynamic error, StackTrace? stackTrace]) {
+    _logger.e(message, error: error, stackTrace: stackTrace);
+  }
+
+  void fatal(dynamic message, [dynamic error, StackTrace? stackTrace]) {
+    _logger.f(message, error: error, stackTrace: stackTrace);
+  }
+
+  void info(dynamic message, [dynamic error, StackTrace? stackTrace]) {
+    _logger.i(message, error: error, stackTrace: stackTrace);
+  }
+
+  void trace(dynamic message, [dynamic error, StackTrace? stackTrace]) {
+    _logger.t(message, error: error, stackTrace: stackTrace);
+  }
+
+  void warning(dynamic message, [dynamic error, StackTrace? stackTrace]) {
+    _logger.w(message, error: error, stackTrace: stackTrace);
+  }
+}