+++ libs/gmana/lib/theme_mode/theme_mode_data.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+class ThemeModeData {
+  final ThemeMode themeMode;
+  final IconData icon;
+  final String label;
+
+  const ThemeModeData({required this.themeMode, required this.icon, required this.label});
+}