+++ libs/gmana/lib/theme_mode/theme_mode_option.dart
@@ -1,0 +1,1 @@
+enum ThemeModeOption { system, dark, light }